# docs/swagger.yaml
basePath: /
definitions:
  requests.Body:
    properties:
      query:
        type: string
    type: object
  responses.Article:
    properties:
      annotation:
        type: string
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
  responses.Error:
    properties:
      message:
        type: string
    type: object
  responses.Response:
    properties:
      articles:
        items:
          $ref: '#/definitions/responses.Article'
        type: array
      found:
        type: integer
    type: object
host: localhost:4000
info:
  contact: {}
  description: Swag documentaion for parser service API
  title: Parser service
  version: 2.0.0
paths:
  /parser/cyberleninka:
    post:
      consumes:
      - application/json
      description: parse one page to get informaion about nr
      parameters:
      - description: create new user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.Body'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Parse one page by current query
      tags:
      - Parser
  /parser/cyberleninka/all:
    post:
      consumes:
      - application/json
      description: parse all site to get informaion about nr
      parameters:
      - description: create new user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.Body'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Parse all pages by current query
      tags:
      - Parser
  /parser/lvrach:
    post:
      consumes:
      - application/json
      description: pars site to get informaion about nr
      parameters:
      - description: create new user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.Body'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Parse page by current query
      tags:
      - Parser
  /parser/lvrach/all:
    post:
      consumes:
      - application/json
      description: pars site to get informaion about nr
      parameters:
      - description: create new user
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/requests.Body'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.Error'
      summary: Parse all pages by current query
      tags:
      - Parser
swagger: "2.0"